requirements.txt
python-telegram-bot
aiohttp_socks
aiohttp
aiosqlite
fastapi
PyYAML